کالجها/دورهٔ جامپ بکاند با جنگو/فروشگاه/requirements.txt
Django>=3.1.2
djangorestframework

